Removed etan test
