--- t0.dir/sbin/.gitignore
@@ -1,2 +1,0 @@
-*.o
-*.d
