--- t0.ok
@@ -1,53 +1,0 @@
-*** 1
-*** init
-*** fd = 3
-*** len = 43
-*** this is nice
-*** we can read and write
-*** close = 0
-*** fd = 3
-*** len = -1
-*** libc.c:27 read error, fd = 3
-*** fd = 100
-*** len = -1
-*** libc.c:27 read error, fd = 100
-*** open again 3
-*** seek 17
-*** in child
-*** argc = 4
-*** argv[0]=shell
-*** argv[1]=a
-*** argv[2]=b
-*** argv[3]=c
-*** we can read and write
-*** back from wait 666
-*** Don't panic
-*** 
-*** 
-***                          oooo$$$$$$$$$$$$oooo
-***                       oo$$$$$$$$$$$$$$$$$$$$$$$$o
-***                    oo$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$o         o$   $$ o$
-***    o $ oo        o$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$o       $$ $$ $$o$
-*** oo $ $ "$      o$$$$$$$$$    $$$$$$$$$$$$$    $$$$$$$$$o       $$$o$$o$
-*** "$$$$$$o$     o$$$$$$$$$      $$$$$$$$$$$      $$$$$$$$$$o    $$$$$$$$
-***   $$$$$$$    $$$$$$$$$$$      $$$$$$$$$$$      $$$$$$$$$$$$$$$$$$$$$$$
-***   $$$$$$$$$$$$$$$$$$$$$$$    $$$$$$$$$$$$$    $$$$$$$$$$$$$$  """$$$
-***    "$$$""""$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$     "$$$
-***     $$$   o$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$     "$$$o
-***    o$$"   $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$       $$$o
-***    $$$    $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$" "$$$$$$ooooo$$$$o
-***   o$$$oooo$$$$$  $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$   o$$$$$$$$$$$$$$$$$
-***   $$$$$$$$"$$$$   $$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$     $$$$""""""""
-***  """"       $$$$    "$$$$$$$$$$$$$$$$$$$$$$$$$$$$"      o$$$
-***             "$$$o     """$$$$$$$$$$$$$$$$$$"$$"         $$$
-***               $$$o          "$$""$$$$$$""""           o$$$
-***                $$$$o                                o$$$"
-***                 "$$$$o      o$$$$$$o"$$$$o        o$$$$
-***                   "$$$$$oo     ""$$$$o$$$$$o   o$$$$""
-***                      ""$$$$$oooo  "$$$o$$$$$$$$$"""
-***                         ""$$$$$$$oo $$$$$$$$$$
-***                                 """"$$$$$$$$$$$
-***                                     $$$$$$$$$$$$
-***                                      $$$$$$$$$$"
-***                                       "$$$""  
-***
