+++ test.dir/etc/panic.txt
