--- test.dir/files/link1
@@ -1,1 +1,0 @@
-msg.txt
