--- test.dir/files/link2
@@ -1,1 +1,0 @@
-link1
