--- test.dir/files/link3
@@ -1,1 +1,0 @@
-/files/msg.txt
