--- test.dir/files/link4
@@ -1,1 +1,0 @@
-link4
