--- test.dir/files/link5
@@ -1,1 +1,0 @@
-../sbin/symlink
