--- test.dir/files/msg.txt
@@ -1,1 +1,0 @@
-I can read from a file!
