+++ test.dir/sbin/crt0.S
@@ -2,10 +2,10 @@
 
 	.global start
 start:
-	.extern printf_init
-	call printf_init
 	.extern heap_init
 	call heap_init
+	.extern printf_init
+	call printf_init
 	call main
 
 	push %eax
