+++ test.dir/sbin/shell
